[CIR][NFC] Refactor ScalarExprEmitter into it's own file
This is more due to a `using namespace clang;` collision.
`CIRGenFunction` has a residual `using namespace` and this header was
importing it incidentally and thus allowed to use non-namespaced decls
from clang. It was either move it to it's own file (where we can define
it in an anonymous namespace) or preserve no usings in headers.