[CIR] Add a helper class, ABIArgInfo, for working with function arguments
This is directly lifted with minimal translation from clang. For the
time being we only really care about two valid types -- Direct (for
simple args such as ints) and Ignore (void).

The rest of the class is to provide the opportunity to assert at places
where clang has differing behavior. e.g. if clang's codegen does
`if (isInAlloca) doSomething();` we want to be able to assert that we
aren't seeing an `isInAlloca` so that we don't forget to handle that
situation later.

This logic is pervasive through many of the following diffs. Most of the
code is redundant and lifted from clang solely so we don't end up with
things going down codepaths we don't anticipate and that we can enforce
it with liberal usages of asserts.

There is also the question of whether or not we need ABIArgInfo at the
moment or anything ABI related at all. We could reason that we should
prefer minimizing ABI interference in the highest levels of CIR.

The hardest part of CIR is going to be getting it correct and working at
all. Minimizing the divergences from clang seems like the easiest way to
do this. If/when we get to the question of making it faster to compile
and better at optimizing code we can focus on luxury things like
lowering ABI decisions to mid-level CIR.