[CIR][NFC] Disable CIRBasedWarnings
Up coming changes to CIRGenerator, CIRGenFunction and CIRGenModule break
this functionality. We'll have to figure out how to rewrire this. We're
going to need to have a system where the CIRBasedWarnings owns the full
CIRGen pipeline.

We're currently just asking CIRGenerator to ask CIRGenModule to build a
function, but given that CGM has state that is required to exist for
CIRGenFunction (which has it's own state) this no longer makes sense.

Simply disable for now.