[CIR] Handle some Parser callbacks from CIRGenerator and CIRGenConsumer
The parser has been calling these member fns without us doing anything
with them. CodeGen requries them to be able to properly handle deferred
decl generation. We'll need to flesh out these implemetnations. But for
now just stub them out and assert that we don't hit a few that we've yet
to see.