[CIR] Add a RAII type for handling inline method definition deferment
Sema calls `HandleInlineFunctionDefinition` which will ping us to
generate a member function definition. But we don't know enough at the
time of invocation to correctly do so. So this mechanism delays
generation via an RAII type that will undo this stack at the end of
HandleTopLevelDecl.

It's currently not complete but later commits will finish it off.