[CIR] Add a mapping from Decl's to cir::Address's for CIRGenFunction
This'll be used for, obviously, getting the previously generated
cir::Address from the Decl. It's currently unused but will receive
usages in a later patch.

This also currently doesn't capture all local variables. Notably it
isn't generating mappings for parameters. But a later patch rebuilds the
parameter generation and thus adding it there would just be redundant.