[CIR][NFC] Add a UnimplementedFeature class to use for guarding against NYIs
Some features in CodeGen aren't really guardable against very well. A
simple example is linkage. We haven't defined anything to do with
linkage yet. So we can't even assert that we don't have a certain type
of it as we do with decl types, paramater count, virtual etc.

So instead introduce a class that explicitly contains a list of static
fns that will return false that you can guard against. If and when a
feature becomes implemented simply changing this return to true will
cause compilation to fail at all the points in which we noted that we
needed to address.

This is a much more explicit way to handle "TODO"s.