[CIR] Remove an unreachable with arrangeCXXMethodDeclaration and impl it
We marked a codepath as unreachable due to not being used before that
will be used coming up.

The fn only checks if we're in an instance function and, if so,
delegates to a currently unreachable'd fn that will do the full chunk of
work with the this ptr available.