[CIR] Add CIRGenRecordLayout type
This is the proxy between the cir::StructType and the AST RecordDecl
that has more information on the type being looked at than is recorded
in the StructType (e.g. the member name).

Currently this is only used as a proxy during codegen, but eventually
we'll be using this type to also build the CIR type with more rich type
information for CIR-level optimizations.