[CIR] Find MLIR as part of the runtimes builds
Clang now needs MLIR and thus a clang distribution also needs MLIR. So
find MLIR first before clang for building the runtimes.

As noted, we'll want the default to be for this to not happen and that
only clangir builds of clang opt into this. e.g. via

```
cmake -DCLANG_BUILD_CLANGIR=ON
```

This would opt into clangir and also propagate it to the runtimes build.