[CIR] Add cir::FuncOp operation
This is necessary to enforce the presence of linkage kind, which
is key to continue implementing methods.

While here add some helpers to apply more reliable parsing for
linkage types.

CodeGen does not use cir::FuncOp just yet, patch a few places that
to keep using mlir::FuncOp for now.