[CIR][CIRGen] Start adding meaningful coroutines tests
Support for coroutines is not complete but we got the minimum support so
we can start to write tests and incrementally cover more functionality.

Here we start by creating allocas for both the task and the promise, with
a first call down to get_return_object in other to get the task.