[CIR][Lowering] Implement a first pass at cir.scope lowering
This implementation mirrors memref::allocascopeop. They don't bother
concerning themselves with whether or not there exists nested scopes,
and I don't think it'll make a big difference for us.

However, they do require a single block in the region which we do not.
So this implemetation might need more work as we add more test cases.