[CIR][CIRGen][Coroutines] Add support for non-void promise calls
LLVM codegen uses a plain rvalue result for coreturn'ing scalars. CIRGen uses
store/load because await_resume is called within cir.await's resume region but
the result is consumed in the outside scope, so we need to dominate all uses.
When lowering to LLVM we can opt out this load/store.