[CIR][CIRGen][Lowering] Support C values unary negation
In C, value negation implicitly casts to int, in C++, it casts to bool.
This patch adds support for the C-specific scenario.

Since both C and C++ share code gen paths for unary negation, the C
cases are handled by casting the boolean type that is already generated
for the C++ cases to an int. The following additions were made:

 - Added C style unary negation test case.
 - Updated CIR cast to support bool_to_int conversions.
 - Lowering bool_to_int conversion as a ZExt operation.

[ghstack-poisoned]