[CIR][CIRGen] Bypass type checking for no-prototype functions
Invalid calls to no-prototype functions should still compile even if
they have undefined behavior. This patch adds a `no_proto` unite attr
to function declarations that bypass argument checking.

The attribute is also used to bypass checks for functions that can
take any number of arguments. This happens when the no-proto function
is externally defined and we can not check its signature.

ghstack-source-id: 94828fc35df62ecfc4eb0b9f0a3ad130e1fb1718
Pull Request resolved: https://github.com/llvm/clangir/pull/130