[CIR][CIRGen][Bugfix] Remove Optional wrapper from MLIR Type
MLIR Types are inherently wrappers. Checking if a value exists is a
matter of checking if the type is null.

The Optional wrapper was causing a bug in the `CIRGenModule::buildGlobal`
method. If an empty type (`mlir::Type{}`) was passed as `Ty`, the
`if (!ty)` clause would evalute to true, because the `!` operator would
simply validate that there was a `mlir::Type` in the optional wrapper.
Then the code gen would continue and try to emit an invalid type.

This patch removes the wrapper, which ensures that the `!Ty` clause now
checks if the given `mlir::Type` is an empty type. In that case, it
generates a real type from the AST Type:

```
  if (!Ty)
    Ty = getTypes().convertTypeForMem(ASTTy);
```

[ghstack-poisoned]