[CIR][CIRGen] Support function basic static variables
Whenever a variable declaration is found, it is created as a global
variable in the module. In C, these variables must be instantiated with
the CIRGenBuilder::createVersionedGlobal to prevent naming conflicts
when multiple static variables are declared across the compilation unit.
In C++, name mangling is used to prevent naming conflicts.