Update base for Update on "[CIR][Lowering] Lower #cir.zero and #cir.null attributes"
LLVM's dialect does not have (at the time of writing) a way to represent
zero-initializers. And, despite being possible to represent null
pointers, it requires it to use a region-based initialization. To avoid
this, LLVM operations that use either of these attributes are marked
with a cir.zero string attribute that will be identified by the
LowerAttrToLLVMIR interface and patch the LLVM IR operation to be
initialized with a zero-initializer or null.

[ghstack-poisoned]