[CIR] Introduce cir-translate to replace mlir-translate (#177)
This change introduces `cir-translate` as a replacement of
`mlir-translate` to convert CIR directly to LLVM IR. The main benefit of
this is to utilize the CIR attribute interface to handle CIR-specific
attributes such as `cir.extra`. Other advantages at this time, besides
the cir attribute support, could be that we can go directly from CIR to
LLVMIR without exposing the intermediate MLIR LLVM dialect form.
Previously `cir-tool` emit the LLVM dialect form and `milr-translate`
took it from there. Now `cir-translate` can directly take CIR and yield
LLVMIR.

I'm also renaming `cir-tool` to `cir-opt` which eventually would be just
a CIR-to-CIR transformer, but for now I'm keeping the functionality of
CIR to LLVM dialect.

So, 
`cir-opt` will do all CIR-to-CIR transforms, just like LLVM `opt` or
`mlir-opt`, and
`cir-translate` will handle CIR to LLVMIR translation, and
LLVMIR-to-LLVMIR transforms, like the LLVM `llc` or the `mlir-translate`