[CIR][CIRGen][Bugfix] Refactor lexical scope locations
Refactors LexicalScopContext to receive a single location and decide
how to use it. This allows the context to have one or two locations for
distinguishing between the start and end of the scope.

Fixes a bug where a mlir::Location was wrongly casted to a mlir::FusedLoc
when building the scope surrounding if statements.

In the rare case where the if statement originates from a preprocessor
macro definition, the location where the macro is expanded is used as
the source location for both the start and the end of the range. Once
these locations are fused, MLIR automatically detects the duplicated
start and end locations and merges them resulting in single a regular
location. This patch allows the if statement to have a single location
as its "begin" and "end" locations.

Note: spelling locations are not tracked due to performance issues.