[CIR][Bugfix] Fix #cir.const_struct parsing
At its core, this patch fixes an issue where the ConstStruct parser
would infer its type from the list of elements and override its explicit
attribute type that had already been parsed. This caused type mismatches
since the name of the StructType would be dropped.

Since `cir.typeinfo` depended on this broken parsing method to work, it
was also fixed to use a `ArrayAttr` instead of a `cir.const_struct` to
store its values.

To simplify parsing and printing struct member on both `cir.typeinfo`
and `cir.const_struct`, the custom `ConstStructMembers` parser/printer
was renamed to `StructMembers` and is now used on both attributes. It
was also refactored to patch malformed spacing between members and to
simplify the code.

[ghstack-poisoned]