[CIR][Bugfix] Fix cir.array getTypeSizeInBits method
Constant initialization of static local arrays would fail due to a
mismatch between the variable and the initializer type size. This patch
fixes the data layout interface implementation for the cir.array type.

A complete array in C/C++ should have its type size in bits equal to
the size of the array times the size of the element type.

ghstack-source-id: 56f3f2918b23309210ad026017bafa37ca03b2d4
Pull Request resolved: https://github.com/llvm/clangir/pull/206