[CIR][ClangTidy] Add CIR support in clang-tidy (to deprecate cir-tidy)
cir-tidy has been our custom hack for using clang-tidy. There's no reason it
needs to be that way - add a CIRModule to clang-tidy and make the lifetime
checker its first check. Some of the way this was built was inspired in
CLANG_TIDY_ENABLE_STATIC_ANALYZER

Once we can fully migrate deps from cir-tidy, we should remove it in favor of
this.