[CIR][Lowering] Implement cir.llvmir.zeroinit operation
Due to the lack of zeroinitializer support in LLVM, some cases are
tricky to lower #cir.zero. An example is when an array is only partially
initialize with #cir.zero attributes. Since we can't just zeroinitialize
the whole array, the current #cir.zero attribute amend does not suffice.
To simplify the lowering, this patch introduces a new operation that is
solely used to generate zeroinitialize LLVM IR constants.

ghstack-source-id: a3fd40ec3ce8970ac4e958076cc17d3fac573696
Pull Request resolved: https://github.com/llvm/clangir/pull/218