[CIR][Codegen][Bugfix] use record layout to generate index for a field (#270)
This is a minor fix similar to the one introduced in #263.
Basically, all calls to the `buildLValueForFieldInitialization` are even
with the origin codegen `emitLValueForFieldInitialization` calls, i.e.
the field index is calculated from the record layout, but not from the
decl `field->getFieldIndex()`.

Added just one test, because looks like we need to implement some `NYI`
features first to test another places e.g. in `CIRGenExprAgg.cpp`,
though I could miss something.

Anyway, given the original codegen doesn't use `getFieldIndex` in these
places, we also should not.

All the remaining usages of `getFieldIndex` are ok.