[CIR][Lowering] Deprecate typed LLVM dialect pointers (#271)
Updates the lowering pass to use only opaque pointers. This essentially
involves updating the type converter to drop pointee types and
explicitly defining the types loaded/stored/GEPed by LLVM operations.

The reasons for this are twofold:

- LLVM dialect is currently transitioning to deprecate typed pointers,
allowing only opaque pointers. The sooner we transition the fewer
changes we will have to make.
- Opaque pointers greatly simplify lowering self-references, since all
self-references in records are wrapped in a pointer.