[CIR] Implement bool->int conversion (#292)
In the CIR CodeGen function `ScalarExprEmitter::buildScalarCast`,
implement conversions from bool to an integral type. This was
inadvertently left out in earlier changes.

Reorganize the code in the function to be more clear, with better
assertion failure messages when encountering an unimplemented construct.

This is a partial fix for issue #290