[CIR] Refactor ScalarExprEmitter::buildScalarCast
Matrix types are already checked for in buildScalarConversion, so they
don't need to be checked for again in buildScalarCast.  Not having to
worry about matrix types means the Element local variables are no longer
necessary.

Remove duplicate code by having a variable to store the CastKind, and
have only one call to Builder.create<CastOp>.

There are no test changes, because this is refactoring only.  There
should be no function changes.