[CIR][CodeGen] Introduce CIRBaseBuilder (#297)
As discussed in #279, we split `CIRGenBuilder` in two parts, which make
some of the helpers usable outside of the `CodeGen` part.
Basically, I placed casts and binary operations into a separate class,
`CIRBaseBuilder`. Later, it can be extended with another helpers. But
right now an idea to have a state less builder as a base one.