[CIR] Fix bool-to-pointer conversions
Conversions from an integer to a pointer are implemented in CIR code gen
as an integral conversion to uintptr_t followed by the
integral-to-pointer conversion.  Conversions from bool to pointer were
following the same code path.  But bool-to-int is a different CastKind
than int-to-int in CIR, and CIR was failing validation.  Fix the integer
to pointer conversion code to correctly handle a source type of bool.

(A conversion from bool to pointer makes no sense and should never
happen in the real world.  But it is legal due to bool being sort of an
integral type.  So we need to support it.)

Also, in ScalarExprEmitter::buildScalarConversion change a couple
not-yet-implemented messages about pointer types into assertion
failures.  Conversions involving pointer types should never go through
ScalarExprEmitter::buildScalarConversion.