[CIR][IR] Refactor parsing/printing of implicitly terminated regions
The `shouldPrintTerm` and `checkBlockTerminator` were replaced in favor
of `omitRegionTerm` and `ensureRegionTerm` respectively. The first is
essentially the same method but simplified. The latter was refactored to
do only two things: check if the terminator omission of a region is
valid and, if so, insert the omitted terminator into the region.

The simplifications mostly leverage the fact that we only omit empty
yield values in a single-block region.

ghstack-source-id: f557556c8f61a86c7edb50c2b2ed4f428be512db
Pull Request resolved: https://github.com/llvm/clangir/pull/310