[CIR][CodeGen] Inline asm: CIR operation (#326)
I will break the PR #308 into pieces and submit them one-by-one.

The first PR introduce CIR operation and the `buildAsmStmt` function.
The latter is the main place for the future changesm and the former was
taken directly from MLIR LLVM IR dialect.
As a result, there is nothing really interesting happen here, but now we
can at least emit cir for an empty inline assembler. And as a first step