[CIR][CodeGen][Lowering] Support multi-block case/default clauses (#356)
This PR adds a support for the multi-block case statements.
Previously, the code example below caused crash in cir verification

Lowering to the `llvm` dialect is pretty straightforward: the same logic
as before is applied to all the region's blocks with no successors, i.e.
we no longer think a case/default region contains only one block.

The `CodeGen` part is a little bit tricky. Previously, any sub-statement
of `case` or`default`, that was not any of them (i.e. neither `case` nor
`default`) was processed with an insertion guard, meaning that the next
sub-statement in the same clause was inserted again in the same block as
the first one. It would be fine, once sub-statement didn't generate any
blocks as well.

For instance,
```
void foo(int a) {
  switch (a)
  {
  case 3:
    return;
    break;
  }
}
```
The `return` statement actually emit a new block after, where the
unreachable code with `break` should be inserted in. That's why we also
need to update `lastCaseBlock` while generating `cir.switch`

This is quite frequent bug in `llvm-test-suite`